# uzum_api_gateway/api/docs/swagger.yaml
definitions:
  catalog_service.Category:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      deleted_at:
        type: integer
      id:
        type: string
      name_en:
        type: string
      name_ru:
        type: string
      name_uz:
        type: string
      order_no:
        type: integer
      parent_id:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  catalog_service.CreateCategory:
    properties:
      active:
        type: boolean
      name_en:
        type: string
      name_ru:
        type: string
      name_uz:
        type: string
      order_no:
        type: integer
      parent_id:
        type: string
      slug:
        type: string
    type: object
...
  catalog_service.Empty:
    type: object
  catalog_service.GetAllCategoryResponse:
    properties:
      Categorys:
        items:
          $ref: '#/definitions/catalog_service.Category'
        type: array
      Count:
        type: integer
    type: object
...
  catalog_service.UpdateCategory:
    properties:
      active:
        type: boolean
      id:
        type: string
      name_en:
        type: string
      name_ru:
        type: string
      name_uz:
        type: string
      order_no:
        type: integer
      parent_id:
        type: string
      slug:
        type: string
    type: object
...
  models.ResponseError:
    properties:
      error: {}
    type: object
  user_service.Branch:
    properties:
      active:
        type: boolean
      address:
        type: string
      close_time:
        type: string
      created_at:
        type: string
      deleted_at:
        type: integer
      id:
        type: string
      location:
        type: string
      name:
        type: string
      open_time:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  user_service.CreateBranch:
    properties:
      active:
        type: boolean
      address:
        type: string
      close_time:
        type: string
      location:
        type: string
      name:
        type: string
      open_time:
        type: string
      phone:
        type: string
    type: object
...
  user_service.Empty3:
    type: object
...
  user_service.GetListBranchResponse:
    properties:
      branches:
        items:
          $ref: '#/definitions/user_service.Branch'
        type: array
      count:
        type: integer
    type: object
...
  user_service.UpdateBranch:
    properties:
      active:
        type: boolean
      address:
        type: string
      close_time:
        type: string
      id:
        type: string
      location:
        type: string
      name:
        type: string
      open_time:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /GetListBranch:
    get:
      consumes:
      - application/json
      description: API for getting list branch
      parameters:
      - description: branches
        in: query
        name: branch
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.GetListBranchResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get List Branch
      tags:
      - branch
...
  /category:
    post:
      consumes:
      - application/json
      description: API for creating a category
      parameters:
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/catalog_service.CreateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.Category'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create a category
      tags:
      - category
  /category/{id}:
    delete:
      consumes:
      - application/json
      description: API for deleting a category by ID
      parameters:
      - description: category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.Empty'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a category by ID
      tags:
      - category
    get:
      consumes:
      - application/json
      description: API for getting a single category by ID
      parameters:
      - description: category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.Category'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a single category by ID
      tags:
      - category
    put:
      consumes:
      - application/json
      description: API for updating a category by ID
      parameters:
      - description: category ID
        in: path
        name: id
        required: true
        type: string
      - description: category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/catalog_service.UpdateCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.Category'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update a category by ID
      tags:
      - category
  /createBranch:
    post:
      consumes:
      - application/json
      description: API for creating branch
      parameters:
      - description: branch
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateBranch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.CreateBranch'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Create branch
      tags:
      - branch
...
  /deleteBranch/{id}:
    delete:
      consumes:
      - application/json
      description: API for deleting a branch by ID
      parameters:
      - description: branch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Empty3'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Delete a branch by ID
      tags:
      - branch
...
  /getallcategory:
    get:
      consumes:
      - application/json
      description: API for getting all categories
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/catalog_service.GetAllCategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get All Categories
      tags:
      - category
...
  /getbyidbranch/{id}:
    get:
      consumes:
      - application/json
      description: API for getting a single branch by ID
      parameters:
      - description: branch ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Branch'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Get a single branch by ID
      tags:
      - branch
...
  /updateBranch/{id}:
    put:
      consumes:
      - application/json
      description: API for updating a branch by ID
      parameters:
      - description: branch ID
        in: path
        name: id
        required: true
        type: string
      - description: branch
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateBranch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Branch'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Update a branch by ID
      tags:
      - branch
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"